XcmPaymentApi::query_weight_to_asset_fee simple common impl (#8281)
# Description

Add a common implementation for
`XcmPaymentApi::query_weight_to_asset_fee` to `pallet-xcm`.

This PR is a simple alternative to #8202 (which could still be useful
for other reasons).
It uses a workaround instead of a big refactoring.

The workaround is:
Computes the weight cost using the provided `WeightTrader`.
This function is supposed to be used ONLY in
`XcmPaymentApi::query_weight_to_asset_fee`
Runtime API implementation, as it can introduce a massive change to the
total issuance.
The provided `WeightTrader` must be the same as the one used in the
XcmExecutor to ensure
uniformity in the weight cost calculation.

NOTE: Currently this function uses a workaround that should be good
enough for all practical
uses: passes `u128::MAX / 2 == 2^127` of the specified asset to the
`WeightTrader` as
payment and computes the weight cost as the difference between this and
the unspent amount.
Some weight traders could add the provided payment to some account's
balance. However,
it should practically never result in overflow because even currencies
with a lot of decimal digits
(say 18) usually have the total issuance of billions (`x * 10^9`) or
trillions (`x * 10^12`) at max,
much less than `2^127 / 10^18 =~ 1.7 * 10^20` (170 billion billion).
Thus, any account's balance
most likely holds less than `2^127`, so adding `2^127` won't result in
`u128` overflow.


## Integration

The Runtime builders can use the `query_weight_to_asset_fee` provided by
`pallet-xcm` in
their XcmPaymentApi implementation.

---------

Co-authored-by: Adrian Catangiu <adrian@parity.io>