collator-protocol-revamp: add PeerManager (#8191)
Implements https://github.com/paritytech/polkadot-sdk/issues/8181
Add peer manager for experimental collator protocol,. The code is
feature-gated by a compile-time feature (which is why I added the silent
label and no prdoc).

PeerManager consists of two components:
- `ConnectedPeers`, which keeps track of currently connected peers as
well as:
    - their reputations for the currently scheduled paras.
    - their protocol versions and the declared paraid.
    - making the decision of whether or not to keep a new connection
- the reputation DB (`Backend` trait implementation), which is left as a
dummy impl, to be implemented in a separate PR:
https://github.com/paritytech/polkadot-sdk/issues/7751

As the wholistic implementation proceeds, the peer manager will likely
suffer some changes that will come in future PRs.
The current state is the best informed guess of all the functionality it
requires, so that the whole implementation is not a big bang and can be
reviewed piece by piece.

TODO:
- [x] add unit tests
- [x] update after merging
https://github.com/paritytech/polkadot-sdk/pull/7955