[AHM] Staking async fixes for XCM and election planning (#8422)
This PR brings a few small fixes related to the XCM messages of
staking-async, among other small fixes:


* [x] Allows `xcm::validate` to check the message size, and we actually
now act upon it in the `staking-async-rc/parachain-runtime`s. The code
is a bit duplicate now, and there is a TOOD about how to better refactor
it later.
* [x] Part of this work is backported separately as
https://github.com/paritytech/polkadot-sdk/pull/8409
* [x] It brings a default `EraElectionPlannerOf` which should be the
right tool to use to ensure elections always happen in time, with an
educated guess based on `ElectionProvider::duration` rather than a
random number.
* [x] It adds a few unit tests about the above
* [x] It silences some logs that were needlessly `INFO`, and makes the
printing of some types a bit more CLI friendly.
* [x] Fix the issue with duplicate voters in solution type: see
https://github.com/paritytech/polkadot-sdk/pull/8585
* [x] Move `PagedRawSolution` to be unbounded, and therefore decode-able
without PoV
* [x] Undo a rename from `ClaimedRewards` to `ErasClaimedRewards`
  * [ ] Needs fixing in westend

---------

Co-authored-by: Adrian Catangiu <adrian@parity.io>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Paolo La Camera <paolo@parity.io>