Introduction of Approval Slashes [Disabling Strategy Stage 4] (#6827)
Implementation of Stage 4 of the new [Disabling
Strategy](https://github.com/paritytech/polkadot-sdk/issues/4359).

Stage 3 needs to be deployed before this goes live. ([Stage 3
PR](https://github.com/paritytech/polkadot-sdk/pull/5724))

Currently only backers are ever slashed as a result of disputes.

This PR creates 3 types of offences with respective slashes:
- ForInvalidBacked (100%) -> hard punishment as backers bear all the
responsbility for validity
- ForInvalidApproved (2%) -> minor punishment to deter lazy validators
- AgainstValid (0%) -> very minor punishment (this causes disablement so
opp cost)

Closes #4746

IMPL details:

This will require a change to parachains primitives which are used both
on the runtime and node side. Rolling it out is time consuming so the
plan is to release it in 2 parts.

First updade runtime only with new offence types but keep old runtime
apis (mapping from new format to old) so the node logic can remain the
same. This will allow us to have the new offences for disputes
concluding on-chain. Past session disputes that require proof
submissions will temporarily not support the new slash types (but they
are rarer). This will make the network more robust in the short term (as
it is much easier to deploy).

As the second part we can update the runtime apis and node side to fully
support the new types.

TODOs:

- [x] new offence types
- [ ] ~~migrations~~ (migration not needed because codec is compatible
in conversion from old->new)
- [ ] ~~migration tests~~
- [x] conversion (new -> old) to preserve old runtime apis
- [x] conversion of disputeProofs (old -> new) to maintain old past
session dispute behaviour
- [ ] new runtime apis (postponed for later)
- [x] disabling for spammy validators zombienet test
- [x] disabling for lazy validators zombienet test
- [ ] unit tests

---------

Co-authored-by: ordian <write@reusable.software>
Co-authored-by: Stephane Gurgenidze <59443568+sw10pa@users.noreply.github.com>
Co-authored-by: ordian <noreply@reusable.software>