Parachains configurations nit comment (#8957)
# Description

I found the previous version of the comment for the
`schedule_config_update` function slightly unclear.

I understood it to mean that the application of two configuration
changes in the same session would lead to the change being applied in
the current session, instead of having to wait until the session change
as usual.

I tried this out on `chopsticks` just to be sure.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>