dispute-coordinator: handle race with offchain disabling (#9050)
Fixes a potential race with off-chain disabling when we learned about
disablement after importing a dispute from that validator.

I think there's no need to handle startup to do deactivation. This will
be only relevant for when a node upgrades to the release with a fix and
writing a migration for that seems like an overkill since this scenario
is very low probability.