bitfield_distribution: fix subsystem clogged at begining of a session (#9094)
`handle_peer_view_change` gets called on NewGossipTopology with the
existing view of the peer to cover for the case when the topology might
arrive late, but in that case in the view will contain old blocks from
previous session, so since the X/Y neighbour change because of the
topology change you end up sending a lot of messages for blocks before
the session changed.

Fix it by checking the send message only for relay chains that are in
the same session as the current topology.

---------

Signed-off-by: Alexandru Gheorghe <alexandru.gheorghe@parity.io>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>