`apply_authorized_force_set_current_code` does not need to consume the whole block (#9202)
There is no need that this dispatchable consumes the full block as this
is just writing the given value to storage. On a chain this is done,
because the runtime changes and thus, a lot of stuff potentially
changes. In the case of the relay chain only on parachain changes and
not the relay chain runtime itself.

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: Adrian Catangiu <adrian@parity.io>