fix: add missing crates bumps and upgrade parity-publish (#9488)
# Description

Adds a few crate bumps associated to PRs which missed to bump them, and
updates parity-publish version across the board to 0.10.6 (to support
rustc 1.88).

Additionally, makes it so that parity-publish-check-compile runs first
on all unreleased prdocs to bump associated crates, and only after
moving those to an `unreleased` directory, runs on the current PR's
prdoc. This is so that we first create a "local release" based on the
unreleased prdocs, and then we follow with a "patch" release based on
the previous local release, considering only the prdoc pushed with the
current PR. If the workflow fails at the end it means current PR missed
certain bumps. If we don't do the plan/apply twice we risk to miss bumps
due to all prdocs being considered (current PR's prdoc + unreleased
ones) when running parity-publish plan/apply, which might result in a
set of crate bumps which are sufficient, but once some unreleased prdocs
will be moved to a stable prodoc directory, because they will be part of
a stable release, then the ones left will not be enough from a bump
perspective (e.g. like it happened in #9320). That's why it is important
to check every PR that adds a prdoc whether it is self-sufficient from a
crates bumping perspective.

If no prdoc is provided, the parity-publish does not need to be taken
into consideration, but it should also pass nonetheless.

## Integration

N/A

## Review Notes

There seems to a be a corner case parity-publish can not easily catch.
All bumps below are a manifestation of it. More details below:

* #8714 - a major bump is necessary for `sp-wasm-interface` - context
here:
https://github.com/paritytech/polkadot-sdk/pull/8714#discussion_r2273355186
* `sp-keystore` was bumped during 2506 in #6010 , and the relevant prdoc
got moved to stable2506 dir in #9320. This moved prdoc coexisted
alongside other unreleased prdocs, and covered a needed patch bump for
`sp-keystore`, that is not easily visible, and also required for crates
publishing IIUC:
1. `sp-io` is major bumped because its direct dependency,
`sp-state-machine`, was major bumped.
2. `sp-io` has a direct dependency on `sp-core` (minor bumped), and
`sp-keystore` (not touched, not bumped by now)
3. `sp-io` fails to compile because it pulls same types from different
`sp-core` versions (it implements `Keystore` trait from `sp-keystore`
with methods signatures referencing types from `sp-core 38.0.0` by using
the `sp-core 0.38.1` - unreleased yet - types, which confuses rustc).
* `sp-rpc` needs a bump too due to pulling `sp-core 38.0.0`, like
`sp-keystore`, and it is an indirect dependency of `polkadot-cli`, which
has also a direct dependency on unreleased `sp-core 38.1.0`, so again,
if we don't bump `sp-rpc` (historically it has been bumped only with
major, but I think we can go with patch on this one), `polkadot-cli`
can't compile.
* `sc-storage-monitor` is in a similar situation as
`sp-rpc`/`sp-keystore` - `polkadot-cli` depends on `sc-storage-monitor`
(which is not bumped, and has a dependency on `sp-core 38.0.0`), but it
also depends on `sp-core 38.1.0`. And yet again, something is used in
`polkadot-cli` from the two different `sp-core` versions, which confuses
rustc.

---------

Signed-off-by: Iulian Barbu <iulian.barbu@parity.io>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Alexander Samusev <41779041+alvicsam@users.noreply.github.com>