Limit the number of signals per XCMP page (#9781)
Right now we have only 2 XCMP signals: `SuspendChannel` and
`ResumeChannel` and we can write at most 1 per page.

Let's also add a limit when reading the signals in a page. Even if now 1
is enough, since in the future we might add more signals, let's have a
limit of 3 per page.

---------

Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>