Rework gas mapping (#9803)
Replacement of https://github.com/paritytech/polkadot-sdk/pull/9740.

Build on top of the new tx payment changes:
https://github.com/paritytech/polkadot-sdk/pull/9780

Started a new PR because due to the rebase on top of the new tx payment
changes this PR is substantially different and I don't want to
invalidate the existing comments on
https://github.com/paritytech/polkadot-sdk/pull/9740 which are not
implemented, yet.

## Overview

This will change the weight to eth fee mapping according to [this
model](https://shade-verse-e97.notion.site/Gas-Mapping-Challenges-Revised-26c8532a7ab580db8222c2ce3023669e).

## Follow ups

This only changes the estimate returned from the dry run and how the
weights are derived from an ethereum transaction. It does not change how
contracts observe the gas. This will be done in a follow up. More
specifically:

1. The `GAS` opcode should return the new gas. As of right now it
returns `ref_time.`
2. The `*_CALL` opcodes should use the passed `gas` parameter and decode
it into `Weight`. As of right now the parameter is ignored.

That said, even without those follow ups this PR should fix all
`InvalidTransaction` errors we are observing.

### Increasing the dimension of `gas_price`

We should add a configurable divisor so that the gas_price is always at
least some gwei. That makes it easier to input the values.

---------

Signed-off-by: Alexander Theißen <alex.theissen@me.com>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Francisco Aguirre <franciscoaguirreperez@gmail.com>
Co-authored-by: PG Herveou <pgherveou@gmail.com>